+++ frontend/src/hooks/useLoadGoals.tsx
@@ -28,6 +28,7 @@ const useLoadGoals = ({formRef, page, setIsLoading, setGoals}: LoadHookParams) =
         const parentId = formData.get('parentId');
         const id = formData.get('idField');
         const locationsWithoutCoords = formData.has('locationsWithoutCoords');
+        const completedGoals = formData.has('completedGoals');
 
         const endpoint = `${REST_ENDPOINT}goals/?page=${page}`;
         let searchFields = '';
@@ -58,6 +59,9 @@ const useLoadGoals = ({formRef, page, setIsLoading, setGoals}: LoadHookParams) =
         }
         if (locationsWithoutCoords === true) {
             searchFields += `&locationsWithoutCoords=true`;
+        }
+        if (completedGoals === true) {
+            searchFields += `&completedGoals=true`;
         }